anistropic filter
include 2 pdes

+++ vol_math_Raw3D_Independt.h
@@ -7,7 +7,7 @@
 #include <iostream>
 using namespace std;
 
-typedef double PIXTYPE;
+typedef float PIXTYPE;
 class Raw2DArray;
 class Raw2D
 {
